+++ lib/utils/currency.value-object.ts
@@ -96,29 +96,60 @@ class CurrencyValueObject extends ValueObject<Prop> {
 	}
 
 	/**
+	 *
+	 * @param value as number
+	 * @returns true if instance value is greater than provided value. else return false.
+	 */
+	isGreaterThan(value: number): boolean {
+		return this.props.value > value;
+	}
+
+	/**
+	 *
+	 * @param value as number
+	 * @returns true if instance value is less than provided value. else return false.
+	 */
+	isLessThan(value: number): boolean {
+		return this.props.value < value;
+	}
+
+	/**
+	 *
+	 * @param value as number
+	 * @returns true if instance value is equal to provided value. else return false.
+	 */
+	isEqualTo(value: number): boolean {
+		return this.props.value === value;
+	}
+
+	/**
 	 * @throws MaxValue `900.719.925.474`
 	 * @event `update` this method may not has effect on the instance value. always check the returned result.
 	 * @description add instance value with provided value
 	 * @param value number to add
 	 * @returns instance of Result type number with calculation result
 	 */
-	add(value: number): Result<number> {
+	add(value: number): Result<CurrencyValueObject> {
 		if (!CurrencyValueObject.isSafeValue(value)) {
-			return Result.fail<number>(
+			return Result.fail<CurrencyValueObject>(
 				`${value} is not a safe number, must be between ${minSafeValue} and ${maxSafeValue}`
 			);
 		}
 		if (this.cents >= maxSafeValue) {
-			return Result.fail<number>('The result is so large for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so large for calculate'
+			);
 		}
 		const valueAsCent = convertValueToCent(value);
 		if (this.cents + valueAsCent >= maxSafeValue) {
-			return Result.fail<number>('The result is so large for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so large for calculate'
+			);
 		}
 
 		this.cents = this.cents + valueAsCent;
 		this.props.value = convertCentToFloat(this.cents);
-		return Result.ok<number>(this.props.value);
+		return Result.ok<CurrencyValueObject>(this);
 	}
 
 	/**
@@ -127,23 +158,27 @@ class CurrencyValueObject extends ValueObject<Prop> {
 	 * @param value number to subtract
 	 * @returns instance of Result type number with calculation result
 	 */
-	subtractBy(value: number): Result<number> {
+	subtractBy(value: number): Result<CurrencyValueObject> {
 		if (!CurrencyValueObject.isSafeValue(value)) {
-			return Result.fail<number>(
+			return Result.fail<CurrencyValueObject>(
 				`${value} is not a safe number, must be between ${minSafeValue} and ${maxSafeValue}`
 			);
 		}
 		if (this.cents <= minSafeValue) {
-			return Result.fail<number>('The result is so large for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so large for calculate'
+			);
 		}
 		const valueAsCent = convertValueToCent(value);
 		if (this.cents - valueAsCent <= minSafeValue) {
-			return Result.fail<number>('The result is so large for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so large for calculate'
+			);
 		}
 
 		this.cents = this.cents - valueAsCent;
 		this.props.value = convertCentToFloat(this.cents);
-		return Result.ok<number>(this.props.value);
+		return Result.ok<CurrencyValueObject>(this);
 	}
 
 	/**
@@ -152,21 +187,25 @@ class CurrencyValueObject extends ValueObject<Prop> {
 	 * @param value number to divide
 	 * @returns instance of Result type number with calculation result
 	 */
-	divideBy(value: number): Result<number> {
+	divideBy(value: number): Result<CurrencyValueObject> {
 		if (!CurrencyValueObject.isSafeValue(value)) {
-			return Result.fail<number>(
+			return Result.fail<CurrencyValueObject>(
 				`${value} is not a safe number, must be between ${minSafeValue} and ${maxSafeValue}`
 			);
 		}
 		if (this.cents <= minSafeValue) {
-			return Result.fail<number>('The result is so small for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so small for calculate'
+			);
 		}
 		if (this.cents / value <= minSafeValue) {
-			return Result.fail<number>('The result is so small for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so small for calculate'
+			);
 		}
 		this.cents = this.cents / value;
 		this.props.value = convertCentToFloat(this.cents);
-		return Result.ok<number>(this.props.value);
+		return Result.ok<CurrencyValueObject>(this);
 	}
 
 	/**
@@ -176,26 +215,30 @@ class CurrencyValueObject extends ValueObject<Prop> {
 	 * @param value number to multiply
 	 * @returns instance of Result type number with calculation result
 	 */
-	multiplyBy(value: number): Result<number> {
+	multiplyBy(value: number): Result<CurrencyValueObject> {
 		if (value < 0.1 && value > -0.1) {
-			return Result.fail<number>(
+			return Result.fail<CurrencyValueObject>(
 				'Invalid value to calculate. Value must be greater than 0.1'
 			);
 		}
 		if (!CurrencyValueObject.isSafeValue(value)) {
-			return Result.fail<number>(
+			return Result.fail<CurrencyValueObject>(
 				`${value} is not a safe number, must be between ${minSafeValue} and ${maxSafeValue}`
 			);
 		}
 		if (this.cents >= maxSafeValue) {
-			return Result.fail<number>('The result is so large for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so large for calculate'
+			);
 		}
 		if (this.cents * value >= maxSafeValue) {
-			return Result.fail<number>('The result is so large for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so large for calculate'
+			);
 		}
 		this.cents = this.cents * value;
 		this.props.value = convertCentToFloat(this.cents);
-		return Result.ok<number>(this.props.value);
+		return Result.ok<CurrencyValueObject>(this);
 	}
 
 	/**
@@ -205,22 +248,26 @@ class CurrencyValueObject extends ValueObject<Prop> {
 	 * @param percent number as percentage to add on instance value
 	 * @returns instance of Result type number with calculation result
 	 */
-	addPercent(percent: number): Result<number> {
+	addPercent(percent: number): Result<CurrencyValueObject> {
 		if (!CurrencyValueObject.isSafeValue(percent)) {
-			return Result.fail<number>(
+			return Result.fail<CurrencyValueObject>(
 				`${percent} is not a safe number, must be between ${minSafeValue} and ${maxSafeValue}`
 			);
 		}
 		if (this.cents >= maxSafeValue) {
-			return Result.fail<number>('The result is so large for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so large for calculate'
+			);
 		}
 		const percentage = calculatePercent(this.cents, percent);
 		if (percentage + this.cents >= maxSafeValue) {
-			return Result.fail<number>('The result is so large for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so large for calculate'
+			);
 		}
 		this.cents = this.cents + percentage;
 		this.props.value = convertCentToFloat(this.cents);
-		return Result.ok<number>(this.props.value);
+		return Result.ok<CurrencyValueObject>(this);
 	}
 
 	/**
@@ -229,22 +276,26 @@ class CurrencyValueObject extends ValueObject<Prop> {
 	 * @param percent number as percentage
 	 * @returns instance of Result type number with calculation result
 	 */
-	subtractPercent(percent: number): Result<number> {
+	subtractPercent(percent: number): Result<CurrencyValueObject> {
 		if (!CurrencyValueObject.isSafeValue(percent)) {
-			return Result.fail<number>(
+			return Result.fail<CurrencyValueObject>(
 				`${percent} is not a safe number, must be between ${minSafeValue} and ${maxSafeValue}`
 			);
 		}
 		if (this.cents <= minSafeValue) {
-			return Result.fail<number>('The result is so large for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so large for calculate'
+			);
 		}
 		const percentage = calculatePercent(this.cents, percent);
 		if (this.cents - percentage <= minSafeValue) {
-			return Result.fail<number>('The result is so large for calculate');
+			return Result.fail<CurrencyValueObject>(
+				'The result is so large for calculate'
+			);
 		}
 		this.cents = this.cents - percentage;
 		this.props.value = convertCentToFloat(this.cents);
-		return Result.ok<number>(this.props.value);
+		return Result.ok<CurrencyValueObject>(this);
 	}
 
 	/**